+++ djangocms_frontend/fields.py
@@ -186,3 +186,11 @@ class AutoNumberInput(forms.NumberInput):  # lgtm [py/missing-call-to-init]
     def __init__(self, *args, **kwargs):
         kwargs.setdefault("attrs", {"class": "auto-field"})
         super().__init__(*args, **kwargs)
+
+
+try:
+    from djangocms_text_ckeditor.fields import HTMLFormField  # noqa
+    HTMLsanitized = True
+except ModuleNotFoundError:
+    HTMLFormField = forms.CharField
+    HTMLsanitized = False