+++ CHANGELOG.rst
@@ -3,11 +3,11 @@ Changelog
 =========
 
 * Add Django 4.0, 4.1 and Python 3.10 support
-* Remove superflous space in some tags
+* Remove superfluous space in some tags
 * Fix figure and blockquote caption
 * Fix HTML injection security bug
 * Removed caption from Image plugin (use figure instead)
-* Fixed js bug which for icon preview
+* Fixed js bug for icon preview
 
 
 1.0.0