+++ CHANGELOG.rst
@@ -2,10 +2,12 @@
 Changelog
 =========
 
-* Remove superflous space in some tags
+* Add Django 4.0, 4.1 and Python 3.10 support
+* Remove superfluous space in some tags
 * Fix figure and blockquote caption
 * Fix HTML injection security bug
 * Removed caption from Image plugin (use figure instead)
+* Fixed js bug for icon preview
 
 
 1.0.0